+++ view/desktop/src/components/Tree/hooks/useMoveTreeNodeEvent.ts
@@ -1,16 +1,7 @@
 import { Dispatch, SetStateAction, useEffect } from "react";
 
-import { Instruction } from "@atlaskit/pragmatic-drag-and-drop-hitbox/dist/types/list-item";
-
 import { MoveNodeEventDetail, TreeNodeProps, TreeProps } from "../types";
-import {
-  addNodeChildrenWithInstruction,
-  addNodeToFolder,
-  findParentNodeByChildUniqueId,
-  hasDescendant,
-  removeNodeFromTree,
-  removeUniqueIdFromTree,
-} from "../utils";
+import { addNodeToTreeWithInstruction, hasDescendant, removeNodeFromTree, removeUniqueIdFromTree } from "../utils";
 
 interface useMoveTreeNodeProps {
   treeId: TreeProps["id"];
@@ -22,61 +13,6 @@ interface useMoveTreeNodeProps {
 
   setTree: Dispatch<SetStateAction<TreeNodeProps>>;
 }
-
-const addNodeToTreeWithInstruction = (
-  tree: TreeNodeProps,
-  targetNode: TreeNodeProps,
-  sourceNode: TreeNodeProps,
-  instruction: Instruction | undefined
-): TreeNodeProps => {
-  const treeWithoutSource = removeNodeFromTree(tree, sourceNode.uniqueId);
-
-  if (!instruction) {
-    if (targetNode.isFolder) {
-      return addNodeToFolder(treeWithoutSource, targetNode.uniqueId, sourceNode);
-    }
-
-    return tree;
-  }
-
-  if (instruction.operation === "combine" && targetNode.isFolder) {
-    return addNodeToFolder(treeWithoutSource, targetNode.uniqueId, sourceNode);
-  }
-
-  const targetParentNode = findParentNodeByChildUniqueId(treeWithoutSource, targetNode.uniqueId);
-  if (!targetParentNode) return treeWithoutSource;
-
-  const indexOfTargetNode = targetParentNode.childNodes.findIndex((child) => child.uniqueId === targetNode.uniqueId);
-  if (indexOfTargetNode === -1) return treeWithoutSource;
-
-  if (instruction.operation === "reorder-before") {
-    return addNodeChildrenWithInstruction(
-      treeWithoutSource,
-      targetParentNode.uniqueId,
-      [
-        ...targetParentNode.childNodes.slice(0, indexOfTargetNode),
-        sourceNode,
-        ...targetParentNode.childNodes.slice(indexOfTargetNode),
-      ],
-      instruction
-    );
-  }
-
-  if (instruction.operation === "reorder-after") {
-    return addNodeChildrenWithInstruction(
-      treeWithoutSource,
-      targetParentNode.uniqueId,
-      [
-        ...targetParentNode.childNodes.slice(0, indexOfTargetNode + 1),
-        sourceNode,
-        ...targetParentNode.childNodes.slice(indexOfTargetNode + 1),
-      ],
-      instruction
-    );
-  }
-
-  return tree;
-};
 
 export const useMoveTreeNodeEvent = ({
   treeId,
