+++ view/desktop/src/components/Tree/utils.ts
@@ -233,6 +233,61 @@ export const getActualDropTargetWithInstruction = (
   };
 };
 
+export const addNodeToTreeWithInstruction = (
+  tree: TreeNodeProps,
+  targetNode: TreeNodeProps,
+  sourceNode: TreeNodeProps,
+  instruction: Instruction | undefined
+): TreeNodeProps => {
+  const treeWithoutSource = removeNodeFromTree(tree, sourceNode.uniqueId);
+
+  if (!instruction) {
+    if (targetNode.isFolder) {
+      return addNodeToFolder(treeWithoutSource, targetNode.uniqueId, sourceNode);
+    }
+
+    return tree;
+  }
+
+  if (instruction.operation === "combine" && targetNode.isFolder) {
+    return addNodeToFolder(treeWithoutSource, targetNode.uniqueId, sourceNode);
+  }
+
+  const targetParentNode = findParentNodeByChildUniqueId(treeWithoutSource, targetNode.uniqueId);
+  if (!targetParentNode) return treeWithoutSource;
+
+  const indexOfTargetNode = targetParentNode.childNodes.findIndex((child) => child.uniqueId === targetNode.uniqueId);
+  if (indexOfTargetNode === -1) return treeWithoutSource;
+
+  if (instruction.operation === "reorder-before") {
+    return addNodeChildrenWithInstruction(
+      treeWithoutSource,
+      targetParentNode.uniqueId,
+      [
+        ...targetParentNode.childNodes.slice(0, indexOfTargetNode),
+        sourceNode,
+        ...targetParentNode.childNodes.slice(indexOfTargetNode),
+      ],
+      instruction
+    );
+  }
+
+  if (instruction.operation === "reorder-after") {
+    return addNodeChildrenWithInstruction(
+      treeWithoutSource,
+      targetParentNode.uniqueId,
+      [
+        ...targetParentNode.childNodes.slice(0, indexOfTargetNode + 1),
+        sourceNode,
+        ...targetParentNode.childNodes.slice(indexOfTargetNode + 1),
+      ],
+      instruction
+    );
+  }
+
+  return tree;
+};
+
 export const canDropNode = (sourceTarget: DropNodeElement, dropTarget: DropNodeElement, node: TreeNodeProps) => {
   if (sourceTarget.node.isFolder === false) {
     // if (hasDirectSimilarDescendant(node, sourceTarget.node)) {
